// src/enums/storageEnum.ts
export enum StorageEnum {
  // token 等信息
  ACCESS_TOKEN_STORE = 'ACCESS_TOKEN',
  // 登录信息
  USER_INFO_STORE = 'LOGIN_INFO',
  // 总需求数
  DEMAND_TOTAL = 'DEMAND_TOTAL',
  DEMAND_TOTAL_YEAR = 'DEMAND_TOTAL_YEAR',
  DEMAND_TOTAL_MONTH = 'DEMAND_TOTAL_MONTH',
  DEMAND_TOTAL_MONTH_PENDING = 'DEMAND_TOTAL_MONTH_PENDING',
  DEMAND_TOTAL_MONTH_UNBEGIN = 'DEMAND_TOTAL_MONTH_UNBEGIN'
}